+++ packages/sdks/angular-sdk/projects/angular-sdk/src/lib/components/applications-portal/applications-portal.component.ts
@@ -20,6 +20,7 @@ export class ApplicationsPortalComponent implements OnInit, OnChanges {
   projectId: string;
   baseUrl?: string;
   baseStaticUrl?: string;
+  baseCdnUrl?: string;
   @Input() widgetId: string;
 
   @Input() theme: 'light' | 'dark' | 'os';
@@ -38,6 +39,7 @@ export class ApplicationsPortalComponent implements OnInit, OnChanges {
     this.projectId = descopeConfig.projectId;
     this.baseUrl = descopeConfig.baseUrl;
     this.baseStaticUrl = descopeConfig.baseStaticUrl;
+    this.baseCdnUrl = descopeConfig.baseCdnUrl;
   }
 
   ngOnInit() {
@@ -57,6 +59,9 @@ export class ApplicationsPortalComponent implements OnInit, OnChanges {
     }
     if (this.baseStaticUrl) {
       this.webComponent.setAttribute('base-static-url', this.baseStaticUrl);
+    }
+    if (this.baseCdnUrl) {
+      this.webComponent.setAttribute('base-cdn-url', this.baseCdnUrl);
     }
     if (this.theme) {
       this.webComponent.setAttribute('theme', this.theme);