+++ packages/sdks/core-js-sdk/test/sdk/saml.test.ts
@@ -48,6 +48,76 @@ describe('saml', () => {
         response: httpResponse,
       });
     });
+
+    it('should return the correct url - sso id', async () => {
+      const httpRespJson = { url: 'http://redirecturl.com/' };
+      const httpResponse = {
+        ok: true,
+        json: () => Promise.resolve(httpRespJson),
+        clone: () => ({
+          json: () => Promise.resolve(httpRespJson),
+        }),
+        status: 200,
+      };
+      mockHttpClient.post.mockResolvedValueOnce(httpResponse);
+      const resp = await sdk.saml.start(
+        'tenant-ID',
+        null,
+        null,
+        null,
+        'some-sso-id',
+      );
+      expect(mockHttpClient.post).toHaveBeenCalledWith(
+        apiPaths.saml.start,
+        {},
+        {
+          queryParams: {
+            tenant: 'tenant-ID',
+            ssoId: 'some-sso-id',
+          },
+        },
+      );
+
+      expect(resp).toEqual({
+        code: 200,
+        data: httpRespJson,
+        ok: true,
+        response: httpResponse,
+      });
+    });
+
+    it('should return the correct url - all params', async () => {
+      const httpRespJson = { url: 'http://redirecturl.com/' };
+      const httpResponse = {
+        ok: true,
+        json: () => Promise.resolve(httpRespJson),
+        clone: () => ({
+          json: () => Promise.resolve(httpRespJson),
+        }),
+        status: 200,
+      };
+      mockHttpClient.post.mockResolvedValueOnce(httpResponse);
+      const resp = await sdk.saml.start('tenant-ID', 'aaa', null, 'ccc', 'ddd');
+      expect(mockHttpClient.post).toHaveBeenCalledWith(
+        apiPaths.saml.start,
+        {},
+        {
+          queryParams: {
+            tenant: 'tenant-ID',
+            redirectURL: 'aaa',
+            ssoId: 'ddd',
+          },
+          token: 'ccc',
+        },
+      );
+
+      expect(resp).toEqual({
+        code: 200,
+        data: httpRespJson,
+        ok: true,
+        response: httpResponse,
+      });
+    });
   });
 
   describe('exchange', () => {