+++ packages/sdks/react-sdk/src/types.ts
@@ -14,7 +14,7 @@ import UserProfileWidget from '@descope/user-profile-widget';
 import ApplicationsPortalWidget from '@descope/applications-portal-widget';
 import createSdk from './sdk';
 
-declare module 'react' {
+declare global {
   namespace JSX {
     interface IntrinsicElements {
       ['descope-wc']: DescopeCustomElement;
@@ -97,6 +97,7 @@ export interface IContext {
   baseUrl?: string;
   styleId?: string;
   baseStaticUrl?: string;
+  baseCdnUrl?: string;
   storeLastAuthenticatedUser?: boolean;
   keepLastAuthenticatedUserAfterLogout?: boolean;
   sdk?: Sdk;
@@ -127,6 +128,16 @@ export type DescopeProps = {
   // use to override client context in flow execution
   client?: Record<string, any>;
   styleId?: string;
+  onScreenUpdate?: (
+    screenName: string,
+    context: Record<string, any>,
+    next: (
+      interactionId: string,
+      form: Record<string, any>,
+    ) => Promise<unknown>,
+    ref: HTMLElement,
+  ) => boolean | Promise<boolean>;
+  children?: React.ReactNode;
 };
 
 export type UserManagementProps = WidgetProps;