+++ packages/widgets/user-management-widget/CHANGELOG.md
@@ -2,6 +2,37 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.6.20](https://github.com/descope/descope-js/compare/user-management-widget-0.6.19...user-management-widget-0.6.20) (2025-02-20)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.24.1`
+## [0.6.19](https://github.com/descope/descope-js/compare/user-management-widget-0.6.18...user-management-widget-0.6.19) (2025-02-12)
+
+### Dependency Updates
+
+* `sdk-mixins` updated to version `0.9.0`
+## [0.6.18](https://github.com/descope/descope-js/compare/user-management-widget-0.6.17...user-management-widget-0.6.18) (2025-02-11)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.24.0`
+## [0.6.17](https://github.com/descope/descope-js/compare/user-management-widget-0.6.16...user-management-widget-0.6.17) (2025-02-11)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.23.10`
+* `sdk-helpers` updated to version `0.2.0`
+* `sdk-mixins` updated to version `0.8.0`
+* `sdk-component-drivers` updated to version `0.2.51`
+## [0.6.16](https://github.com/descope/descope-js/compare/user-management-widget-0.6.15...user-management-widget-0.6.16) (2025-02-11)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.23.9`
+* `sdk-helpers` updated to version `0.1.67`
+* `sdk-mixins` updated to version `0.7.0`
+* `sdk-component-drivers` updated to version `0.2.50`
 ## [0.6.15](https://github.com/descope/descope-js/compare/user-management-widget-0.6.14...user-management-widget-0.6.15) (2025-02-02)
 
 ### Dependency Updates